pyproject.toml
[project]
name = "ppbase"
version = "0.1.6"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "pyjwt>=2.9.0",
    "passlib[bcrypt]>=1.7.4",
    "lark>=1.2.0",
    "python-multipart>=0.0.12",
    "aiofiles>=24.1.0",
    "boto3>=1.35.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "testcontainers[postgres]>=4.0",
]

[project.scripts]
ppbase = "ppbase.__main__:main"

[tool.setuptools.packages.find]
include = ["ppbase*"]

[tool.setuptools.package-data]
"ppbase" = ["admin/dist/**/*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
testpaths = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
